Resolve existing build failure
