+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/tests/Azure.Monitor.OpenTelemetry.Exporter.Tests/CommonTestFramework/MockTransmitter.cs
@@ -9,14 +9,13 @@ using System.Threading.Tasks;
 using Azure.Monitor.OpenTelemetry.Exporter.Internals;
 using Azure.Monitor.OpenTelemetry.Exporter.Internals.CustomerSdkStats;
 using Azure.Monitor.OpenTelemetry.Exporter.Internals.Diagnostics;
-using Azure.Monitor.OpenTelemetry.Exporter.Internals.MultiTenant;
 using Azure.Monitor.OpenTelemetry.Exporter.Models;
 
 using OpenTelemetry;
 
 namespace Azure.Monitor.OpenTelemetry.Exporter.Tests.CommonTestFramework
 {
-    internal class MockTransmitter : ITransmitter, IMultiTenantTransmitter
+    internal partial class MockTransmitter : ITransmitter
     {
         public readonly IList<TelemetryItem> TelemetryItems;
 
@@ -43,27 +42,6 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Tests.CommonTestFramework
         }
 
         public int TrackAsyncCallCount { get; private set; }
-
-        /// <summary>
-        /// Per multi-tenant send: the destination endpoint and the items delivered to it.
-        /// </summary>
-        public readonly List<(string IngestionEndpoint, TelemetryItem[] TelemetryItems)> Sends = new();
-
-        public ExportResult MultiTenantResult { get; set; } = ExportResult.Success;
-
-        public ExportResult Track(EndpointRouteBatch routeBatch, TelemetryItemOrigin origin, CancellationToken cancellationToken)
-        {
-            lock (this.Sends)
-            {
-                for (int i = 0; i < routeBatch.Count; i++)
-                {
-                    var group = routeBatch[i];
-                    Sends.Add((group.IngestionEndpoint, group.TelemetryItems.ToArray()));
-                }
-            }
-
-            return MultiTenantResult;
-        }
 
         public ValueTask TransmitFromStorage(long maxFileToTransmit, bool async, CancellationToken cancellationToken)
         {
